Refactored files to organise the files better
