+++ app/src/main/AndroidManifest.xml
@@ -28,7 +28,7 @@
                 android:value="com.lxdnz.nz.movieproject.MainActivity" />
         </activity>
         <activity
-            android:name=".SettingsActivity"
+            android:name=".preferenceActivity.SettingsActivity"
             android:label="@string/title_activity_settings"
             android:parentActivityName=".MainActivity">
             <meta-data
