+++ app/src/main/java/com/lxdnz/nz/movieproject/MainActivity.java
@@ -1,6 +1,5 @@
 package com.lxdnz.nz.movieproject;
 
-import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
@@ -9,6 +8,8 @@ import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
 import android.view.Menu;
 import android.view.MenuItem;
+
+import com.lxdnz.nz.movieproject.preferenceActivity.SettingsActivity;
 
 public class MainActivity extends AppCompatActivity {
