+++ app/src/main/java/com/lxdnz/nz/movieproject/preferenceActivity/AppCompatPreferenceActivity.java
@@ -1,4 +1,4 @@
-package com.lxdnz.nz.movieproject;
+package com.lxdnz.nz.movieproject.preferenceActivity;
 
 import android.content.res.Configuration;
 import android.os.Bundle;
